mysite/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
crispy-bootstrap5==0.7
Django==4.1.2
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-widget-tweaks==1.4.12
Pillow==9.3.0
python-dotenv==0.21.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.6
